+++ gfx/vr/gfxVR.h
@@ -201,14 +201,27 @@ protected:
   nsCOMPtr<nsIScreen> mScreen;
 };
 
-class VRHMDManagerOculusImpl;
-class VRHMDManagerOculus {
-  static VRHMDManagerOculusImpl *mImpl;
+class VRHMDManager {
 public:
-  static bool PlatformInit();
-  static bool Init();
-  static void Destroy();
-  static void GetOculusHMDs(nsTArray<nsRefPtr<VRHMDInfo> >& aHMDResult);
+  static void ManagerInit();
+  static void ManagerDestroy();
+  static void GetAllHMDs(nsTArray<nsRefPtr<VRHMDInfo>>& aHMDResult);
+
+protected:
+  typedef nsTArray<nsRefPtr<VRHMDManager>> VRHMDManagerArray;
+  static VRHMDManagerArray *sManagers;
+
+public:
+  NS_INLINE_DECL_THREADSAFE_REFCOUNTING(VRHMDManager)
+
+  virtual bool PlatformInit() = 0;
+  virtual bool Init() = 0;
+  virtual void Destroy() = 0;
+  virtual void GetHMDs(nsTArray<nsRefPtr<VRHMDInfo>>& aHMDResult) = 0;
+
+protected:
+  VRHMDManager() { }
+  virtual ~VRHMDManager() { }
 };
 
 } // namespace gfx
